+++ R/step-subset.R
@@ -178,6 +178,37 @@ transmute.dtplyr_step <- function(.data, ...) {
   step_subset_j(.data, vars = names(dots), j = j)
 }
 
+#' @importFrom dplyr count
+#' @export
+count.dtplyr_step <- function(.data, ..., wt = NULL, sort = FALSE, name = NULL) {
+  if (!missing(...)) {
+    out <- group_by(.data, ..., .add = TRUE)
+  } else {
+    out <- .data
+  }
+
+  wt <- enexpr(wt)
+  if (is.null(wt)) {
+    n <- expr(n())
+  } else {
+    n <- expr(sum(!!wt, na.rm = TRUE))
+  }
+
+  if (is.null(name)) {
+    name <- "n"
+  } else if (!is_string(name)) {
+    abort("`name` must be a string")
+  }
+
+  out <- summarise(out, !!name := !!n)
+
+  if (sort) {
+    out <- arrange(out, desc(!!sym(name)))
+  }
+
+  out
+}
+
 # exported onLoad
 filter.dtplyr_step <- function(.data, ...) {
   dots <- capture_dots(.data, ..., .j = FALSE)